requirements.txt
albumentations==0.4.3
backcall==0.1.0
certifi==2019.11.28
cffi==1.10.0
chardet==3.0.4
cupy-cuda80==7.1.1
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
docopt==0.6.2
fastrlock==0.4
fire==0.2.1
idna==2.8
imageio==2.6.1
imgaug==0.2.6
ipdb==0.12.3
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
kiwisolver==1.1.0
matplotlib==3.1.2
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mlconfig==0.0.4
networkx==2.4
nose==1.3.7
numpy==1.18.1
olefile==0.44
opencv-python==4.1.1.26
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
pipreqs==0.4.10
pprint==0.1
prompt-toolkit==3.0.3
protobuf==3.11.2
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.10.0
sklearn==0.0
tensorboardX==2.0
termcolor==1.1.0
torch==1.0.0
torch-encoding==1.0.1
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.2.1
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
urllib3==1.25.8
visdom==0.1.8.9
wcwidth==0.1.8
websocket-client==0.57.0
websockets==8.1
yacs==0.1.6
yarg==0.1.9
